Moved heatmap series into a subfolder.
