+++ ts/Series/Heatmap/HeatmapSeries.ts
@@ -7,38 +7,111 @@
  *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
  *
  * */
+
 'use strict';
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = function (d, b) {
-        extendStatics = Object.setPrototypeOf ||
-            ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-            function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-        return extendStatics(d, b);
-    };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
-import BaseSeries from '../Core/Series/Series.js';
-var _a = BaseSeries.seriesTypes, ColumnSeries = _a.column, ScatterSeries = _a.scatter;
-import ColorMapMixin from '../Mixins/ColorMapSeries.js';
-var colorMapPointMixin = ColorMapMixin.colorMapPointMixin, colorMapSeriesMixin = ColorMapMixin.colorMapSeriesMixin;
-import H from '../Core/Globals.js';
-var noop = H.noop;
-import LegendSymbolMixin from '../Mixins/LegendSymbol.js';
-import LineSeries from '../Series/Line/LineSeries.js';
-import palette from '../Core/Color/Palette.js';
-import SVGRenderer from '../Core/Renderer/SVG/SVGRenderer.js';
-var symbols = SVGRenderer.prototype.symbols;
-import U from '../Core/Utilities.js';
-var clamp = U.clamp, extend = U.extend, fireEvent = U.fireEvent, isNumber = U.isNumber, merge = U.merge, pick = U.pick;
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type AnimationOptionsObject from '../../Core/Animation/AnimationOptionsObject';
+import type ColorAxis from '../../Core/Axis/ColorAxis';
+import type ColorType from '../../Core/Color/ColorType';
+import type DataExtremesObject from '../../Core/Series/DataExtremesObject';
+import type Point from '../../Core/Series/Point.js';
+import type { PointStateHoverOptions } from '../../Core/Series/PointOptions';
+import type ScatterPointOptions from '../Scatter/ScatterPointOptions';
+import type ScatterSeriesOptions from '../Scatter/ScatterSeriesOptions';
+import type { SeriesStatesOptions } from '../../Core/Series/SeriesOptions';
+import type { StatesOptionsKey } from '../../Core/Series/StatesOptions';
+import type SVGAttributes from '../../Core/Renderer/SVG/SVGAttributes';
+import type SVGPath from '../../Core/Renderer/SVG/SVGPath';
+import BaseSeries from '../../Core/Series/Series.js';
+const {
+    seriesTypes: {
+        column: ColumnSeries,
+        scatter: ScatterSeries
+    }
+} = BaseSeries;
+import ColorMapMixin from '../../Mixins/ColorMapSeries.js';
+const {
+    colorMapPointMixin,
+    colorMapSeriesMixin
+} = ColorMapMixin;
+import H from '../../Core/Globals.js';
+const { noop } = H;
+import LegendSymbolMixin from '../../Mixins/LegendSymbol.js';
+import LineSeries from '../../Series/Line/LineSeries.js';
+import palette from '../../Core/Color/Palette.js';
+import SVGRenderer from '../../Core/Renderer/SVG/SVGRenderer.js';
+const {
+    prototype: {
+        symbols
+    }
+} = SVGRenderer;
+import U from '../../Core/Utilities.js';
+const {
+    clamp,
+    extend,
+    fireEvent,
+    isNumber,
+    merge,
+    pick
+} = U;
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+declare module '../../Core/Series/SeriesLike' {
+    interface SeriesLike {
+        valueMax?: number;
+        valueMin?: number;
+    }
+}
+
+declare module '../../Core/Series/SeriesOptions' {
+    interface SeriesStateHoverOptions {
+        brightness?: number;
+    }
+}
+
+/**
+ * Internal types
+ * @private
+ */
+declare global {
+    namespace Highcharts {
+        interface HeatmapPointCellAttributes extends Record<string, number> {
+            x1: number;
+            x2: number;
+            y1: number;
+            y2: number;
+        }
+        interface HeatmapPointOptions extends ScatterPointOptions {
+            pointPadding?: HeatmapPoint['pointPadding'];
+            value?: HeatmapPoint['value'];
+        }
+        interface HeatmapSeriesOptions extends ScatterSeriesOptions {
+            colsize?: number;
+            nullColor?: ColorType;
+            pointPadding?: HeatmapPoint['pointPadding'];
+            rowsize?: number;
+            states?: SeriesStatesOptions<HeatmapSeries>;
+        }
+    }
+}
+
 /* *
  *
  *  Class
  *
  * */
+
 /**
  * @private
  * @class
@@ -46,223 +119,14 @@ var clamp = U.clamp, extend = U.extend, fireEvent = U.fireEvent, isNumber = U.is
  *
  * @augments Highcharts.Series
  */
-var HeatmapSeries = /** @class */ (function (_super) {
-    __extends(HeatmapSeries, _super);
-    function HeatmapSeries() {
-        /* *
-         *
-         *  Static Properties
-         *
-         * */
-        var _this = _super !== null && _super.apply(this, arguments) || this;
-        /* *
-         *
-         *  Properties
-         *
-         * */
-        _this.colorAxis = void 0;
-        _this.data = void 0;
-        _this.options = void 0;
-        _this.points = void 0;
-        _this.valueMax = NaN;
-        _this.valueMin = NaN;
-        return _this;
-        /* eslint-enable valid-jsdoc */
-    }
+class HeatmapSeries extends ScatterSeries {
+
     /* *
      *
-     *  Functions
+     *  Static Properties
      *
      * */
-    /* eslint-disable valid-jsdoc */
-    /**
-     * @private
-     */
-    HeatmapSeries.prototype.drawPoints = function () {
-        var _this = this;
-        // In styled mode, use CSS, otherwise the fill used in the style
-        // sheet will take precedence over the fill attribute.
-        var seriesMarkerOptions = this.options.marker || {};
-        if (seriesMarkerOptions.enabled || this._hasPointMarkers) {
-            LineSeries.prototype.drawPoints.call(this);
-            this.points.forEach(function (point) {
-                point.graphic &&
-                    point.graphic[_this.chart.styledMode ? 'css' : 'animate'](_this.colorAttribs(point));
-            });
-        }
-    };
-    /**
-     * @private
-     */
-    HeatmapSeries.prototype.getExtremes = function () {
-        // Get the extremes from the value data
-        var _a = LineSeries.prototype.getExtremes
-            .call(this, this.valueData), dataMin = _a.dataMin, dataMax = _a.dataMax;
-        if (isNumber(dataMin)) {
-            this.valueMin = dataMin;
-        }
-        if (isNumber(dataMax)) {
-            this.valueMax = dataMax;
-        }
-        // Get the extremes from the y data
-        return LineSeries.prototype.getExtremes.call(this);
-    };
-    /**
-     * Override to also allow null points, used when building the k-d-tree for
-     * tooltips in boost mode.
-     * @private
-     */
-    HeatmapSeries.prototype.getValidPoints = function (points, insideOnly) {
-        return LineSeries.prototype.getValidPoints.call(this, points, insideOnly, true);
-    };
-    /**
-     * Define hasData function for non-cartesian series. Returns true if the
-     * series has points at all.
-     * @private
-     */
-    HeatmapSeries.prototype.hasData = function () {
-        return !!this.processedXData.length; // != 0
-    };
-    /**
-     * Override the init method to add point ranges on both axes.
-     * @private
-     */
-    HeatmapSeries.prototype.init = function () {
-        var options;
-        LineSeries.prototype.init.apply(this, arguments);
-        options = this.options;
-        // #3758, prevent resetting in setData
-        options.pointRange = pick(options.pointRange, options.colsize || 1);
-        // general point range
-        this.yAxis.axisPointRange = options.rowsize || 1;
-        // Bind new symbol names
-        extend(symbols, {
-            ellipse: symbols.circle,
-            rect: symbols.square
-        });
-    };
-    /**
-     * @private
-     */
-    HeatmapSeries.prototype.markerAttribs = function (point, state) {
-        var pointMarkerOptions = point.marker || {}, seriesMarkerOptions = this.options.marker || {}, seriesStateOptions, pointStateOptions, shapeArgs = point.shapeArgs || {}, hasImage = point.hasImage, attribs = {};
-        if (hasImage) {
-            return {
-                x: point.plotX,
-                y: point.plotY
-            };
-        }
-        // Setting width and height attributes on image does not affect
-        // on its dimensions.
-        if (state) {
-            seriesStateOptions = seriesMarkerOptions.states[state] || {};
-            pointStateOptions = pointMarkerOptions.states &&
-                pointMarkerOptions.states[state] || {};
-            [['width', 'x'], ['height', 'y']].forEach(function (dimension) {
-                // Set new width and height basing on state options.
-                attribs[dimension[0]] = (pointStateOptions[dimension[0]] ||
-                    seriesStateOptions[dimension[0]] ||
-                    shapeArgs[dimension[0]]) + (pointStateOptions[dimension[0] + 'Plus'] ||
-                    seriesStateOptions[dimension[0] + 'Plus'] || 0);
-                // Align marker by a new size.
-                attribs[dimension[1]] = shapeArgs[dimension[1]] +
-                    (shapeArgs[dimension[0]] - attribs[dimension[0]]) / 2;
-            });
-        }
-        return state ? attribs : shapeArgs;
-    };
-    /**
-     * @private
-     */
-    HeatmapSeries.prototype.pointAttribs = function (point, state) {
-        var series = this, attr = LineSeries.prototype.pointAttribs.call(series, point, state), seriesOptions = series.options || {}, plotOptions = series.chart.options.plotOptions || {}, seriesPlotOptions = plotOptions.series || {}, heatmapPlotOptions = plotOptions.heatmap || {}, stateOptions, brightness, 
-        // Get old properties in order to keep backward compatibility
-        borderColor = seriesOptions.borderColor ||
-            heatmapPlotOptions.borderColor ||
-            seriesPlotOptions.borderColor, borderWidth = seriesOptions.borderWidth ||
-            heatmapPlotOptions.borderWidth ||
-            seriesPlotOptions.borderWidth ||
-            attr['stroke-width'];
-        // Apply lineColor, or set it to default series color.
-        attr.stroke = ((point && point.marker && point.marker.lineColor) ||
-            (seriesOptions.marker && seriesOptions.marker.lineColor) ||
-            borderColor ||
-            this.color);
-        // Apply old borderWidth property if exists.
-        attr['stroke-width'] = borderWidth;
-        if (state) {
-            stateOptions =
-                merge(seriesOptions.states[state], seriesOptions.marker &&
-                    seriesOptions.marker.states[state], point &&
-                    point.options.states &&
-                    point.options.states[state] || {});
-            brightness = stateOptions.brightness;
-            attr.fill =
-                stateOptions.color ||
-                    H.color(attr.fill).brighten(brightness || 0).get();
-            attr.stroke = stateOptions.lineColor;
-        }
-        return attr;
-    };
-    /**
-     * @private
-     */
-    HeatmapSeries.prototype.setClip = function (animation) {
-        var series = this, chart = series.chart;
-        LineSeries.prototype.setClip.apply(series, arguments);
-        if (series.options.clip !== false || animation) {
-            series.markerGroup
-                .clip((animation || series.clipBox) && series.sharedClipKey ?
-                chart[series.sharedClipKey] :
-                chart.clipRect);
-        }
-    };
-    /**
-     * @private
-     */
-    HeatmapSeries.prototype.translate = function () {
-        var series = this, options = series.options, symbol = options.marker && options.marker.symbol || '', shape = symbols[symbol] ? symbol : 'rect', options = series.options, hasRegularShape = ['circle', 'square'].indexOf(shape) !== -1;
-        series.generatePoints();
-        series.points.forEach(function (point) {
-            var pointAttr, sizeDiff, hasImage, cellAttr = point.getCellAttributes(), shapeArgs = {
-                x: Math.min(cellAttr.x1, cellAttr.x2),
-                y: Math.min(cellAttr.y1, cellAttr.y2),
-                width: Math.max(Math.abs(cellAttr.x2 - cellAttr.x1), 0),
-                height: Math.max(Math.abs(cellAttr.y2 - cellAttr.y1), 0)
-            };
-            hasImage = point.hasImage =
-                (point.marker && point.marker.symbol || symbol || '')
-                    .indexOf('url') === 0;
-            // If marker shape is regular (symetric), find shorter
-            // cell's side.
-            if (hasRegularShape) {
-                sizeDiff = Math.abs(shapeArgs.width - shapeArgs.height);
-                shapeArgs.x = Math.min(cellAttr.x1, cellAttr.x2) +
-                    (shapeArgs.width < shapeArgs.height ? 0 : sizeDiff / 2);
-                shapeArgs.y = Math.min(cellAttr.y1, cellAttr.y2) +
-                    (shapeArgs.width < shapeArgs.height ? sizeDiff / 2 : 0);
-                shapeArgs.width = shapeArgs.height =
-                    Math.min(shapeArgs.width, shapeArgs.height);
-            }
-            pointAttr = {
-                plotX: (cellAttr.x1 + cellAttr.x2) / 2,
-                plotY: (cellAttr.y1 + cellAttr.y2) / 2,
-                clientX: (cellAttr.x1 + cellAttr.x2) / 2,
-                shapeType: 'path',
-                shapeArgs: merge(true, shapeArgs, {
-                    d: symbols[shape](shapeArgs.x, shapeArgs.y, shapeArgs.width, shapeArgs.height)
-                })
-            };
-            if (hasImage) {
-                point.marker = {
-                    width: shapeArgs.width,
-                    height: shapeArgs.height
-                };
-            }
-            extend(point, pointAttr);
-        });
-        fireEvent(series, 'afterTranslate');
-    };
+
     /**
      * A heatmap is a graphical representation of data where the individual
      * values contained in a matrix are represented as colors.
@@ -284,15 +148,18 @@ var HeatmapSeries = /** @class */ (function (_super) {
      * @product      highcharts highmaps
      * @optionparent plotOptions.heatmap
      */
-    HeatmapSeries.defaultOptions = merge(ScatterSeries.defaultOptions, {
+    public static defaultOptions: Highcharts.HeatmapSeriesOptions = merge(ScatterSeries.defaultOptions, {
+
         /**
          * Animation is disabled by default on the heatmap series.
          */
         animation: false,
+
         /**
          * The border width for each heat map item.
          */
         borderWidth: 0,
+
         /**
          * Padding between the points in the heatmap.
          *
@@ -301,10 +168,12 @@ var HeatmapSeries = /** @class */ (function (_super) {
          * @since     6.0
          * @apioption plotOptions.heatmap.pointPadding
          */
+
         /**
          * @default   value
          * @apioption plotOptions.heatmap.colorKey
          */
+
         /**
          * The main color of the series. In heat maps this color is rarely used,
          * as we mostly use the color to denote the value of each point. Unless
@@ -316,6 +185,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
          * @product   highcharts
          * @apioption plotOptions.heatmap.color
          */
+
         /**
          * The column size - how many X axis units each column in the heatmap
          * should span.
@@ -331,6 +201,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
          * @product   highcharts highmaps
          * @apioption plotOptions.heatmap.colsize
          */
+
         /**
          * The row size - how many Y axis units each heatmap row should span.
          *
@@ -345,6 +216,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
          * @product   highcharts highmaps
          * @apioption plotOptions.heatmap.rowsize
          */
+
         /**
          * The color applied to null points. In styled mode, a general CSS class
          * is applied instead.
@@ -352,14 +224,15 @@ var HeatmapSeries = /** @class */ (function (_super) {
          * @type {Highcharts.ColorString|Highcharts.GradientColorObject|Highcharts.PatternObject}
          */
         nullColor: palette.neutralColor3,
+
         dataLabels: {
-            formatter: function () {
-                return this.point.value;
+            formatter: function (): (number|null) { // #2945
+                return (this.point as HeatmapPoint).value;
             },
             inside: true,
             verticalAlign: 'middle',
             crop: false,
-            overflow: false,
+            overflow: false as any,
             padding: 0 // #3837
         },
         /**
@@ -407,6 +280,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
                      * @product   highcharts highmaps
                      * @apioption plotOptions.heatmap.marker.states.hover.width
                      */
+
                     /**
                      * Set the marker's fixed height on hover state.
                      *
@@ -418,6 +292,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
                      * @product   highcharts highmaps
                      * @apioption plotOptions.heatmap.marker.states.hover.height
                      */
+
                     /**
                      * The number of pixels to increase the width of the
                      * selected point.
@@ -430,6 +305,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
                      * @product   highcharts highmaps
                      * @apioption plotOptions.heatmap.marker.states.hover.widthPlus
                      */
+
                     /**
                      * The number of pixels to increase the height of the
                      * selected point.
@@ -442,6 +318,7 @@ var HeatmapSeries = /** @class */ (function (_super) {
                      * @product   highcharts highmaps
                      * @apioption plotOptions.heatmap.marker.states.hover.heightPlus
                      */
+
                     /**
                      * The additional line width for a hovered point.
                      *
@@ -456,65 +333,75 @@ var HeatmapSeries = /** @class */ (function (_super) {
                  * @excluding radius
                  */
                 select: {
-                /**
-                 * Set the marker's fixed width on select state.
-                 *
-                 * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-width
-                 *         70px fixed marker's width and height on hover
-                 *
-                 * @type      {number|undefined}
-                 * @default   undefined
-                 * @product   highcharts highmaps
-                 * @apioption plotOptions.heatmap.marker.states.select.width
-                 */
-                /**
-                 * Set the marker's fixed height on select state.
-                 *
-                 * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-width
-                 *         70px fixed marker's width and height on hover
-                 *
-                 * @type      {number|undefined}
-                 * @default   undefined
-                 * @product   highcharts highmaps
-                 * @apioption plotOptions.heatmap.marker.states.select.height
-                 */
-                /**
-                 * The number of pixels to increase the width of the
-                 * selected point.
-                 *
-                 * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-widthplus
-                 *         20px greater width and height on hover
-                 *
-                 * @type      {number|undefined}
-                 * @default   undefined
-                 * @product   highcharts highmaps
-                 * @apioption plotOptions.heatmap.marker.states.select.widthPlus
-                 */
-                /**
-                 * The number of pixels to increase the height of the
-                 * selected point.
-                 *
-                 * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-widthplus
-                 *         20px greater width and height on hover
-                 *
-                 * @type      {number|undefined}
-                 * @default   undefined
-                 * @product   highcharts highmaps
-                 * @apioption plotOptions.heatmap.marker.states.select.heightPlus
-                 */
+                    /**
+                     * Set the marker's fixed width on select state.
+                     *
+                     * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-width
+                     *         70px fixed marker's width and height on hover
+                     *
+                     * @type      {number|undefined}
+                     * @default   undefined
+                     * @product   highcharts highmaps
+                     * @apioption plotOptions.heatmap.marker.states.select.width
+                     */
+
+                    /**
+                     * Set the marker's fixed height on select state.
+                     *
+                     * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-width
+                     *         70px fixed marker's width and height on hover
+                     *
+                     * @type      {number|undefined}
+                     * @default   undefined
+                     * @product   highcharts highmaps
+                     * @apioption plotOptions.heatmap.marker.states.select.height
+                     */
+
+                    /**
+                     * The number of pixels to increase the width of the
+                     * selected point.
+                     *
+                     * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-widthplus
+                     *         20px greater width and height on hover
+                     *
+                     * @type      {number|undefined}
+                     * @default   undefined
+                     * @product   highcharts highmaps
+                     * @apioption plotOptions.heatmap.marker.states.select.widthPlus
+                     */
+
+                    /**
+                     * The number of pixels to increase the height of the
+                     * selected point.
+                     *
+                     * @sample {highcharts} maps/plotoptions/heatmap-marker-states-hover-widthplus
+                     *         20px greater width and height on hover
+                     *
+                     * @type      {number|undefined}
+                     * @default   undefined
+                     * @product   highcharts highmaps
+                     * @apioption plotOptions.heatmap.marker.states.select.heightPlus
+                     */
                 }
             }
         },
+
         clip: true,
+
         /** @ignore-option */
-        pointRange: null,
+        pointRange: null, // dynamically set to colsize by default
+
         tooltip: {
             pointFormat: '{point.x}, {point.y}: {point.value}<br/>'
         },
+
         states: {
+
             hover: {
+
                 /** @ignore-option */
-                halo: false,
+                halo: false, // #3406, halo is disabled on heatmaps by default
+
                 /**
                  * How much to brighten the point on interaction. Requires the
                  * main color to be defined in hex or rgb(a) format.
@@ -525,98 +412,523 @@ var HeatmapSeries = /** @class */ (function (_super) {
                  */
                 brightness: 0.2
             }
+
         }
-    });
-    return HeatmapSeries;
-}(ScatterSeries));
-extend(HeatmapSeries.prototype, {
-    /**
-     * @private
-     */
-    alignDataLabel: ColumnSeries.prototype.alignDataLabel,
-    axisTypes: colorMapSeriesMixin.axisTypes,
-    colorAttribs: colorMapSeriesMixin.colorAttribs,
-    colorKey: colorMapSeriesMixin.colorKey,
-    directTouch: true,
-    /**
-     * @private
-     */
-    drawLegendSymbol: LegendSymbolMixin.drawRectangle,
-    hasPointSpecificOptions: true,
-    /**
-     * @ignore
-     * @deprecated
-     */
-    getBox: noop,
-    getExtremesFromAll: true,
-    getSymbol: LineSeries.prototype.getSymbol,
-    parallelArrays: colorMapSeriesMixin.parallelArrays,
-    pointArrayMap: ['y', 'value'],
-    trackerGroups: colorMapSeriesMixin.trackerGroups
-});
-/* *
- *
- *  Class
- *
- * */
-var HeatmapPoint = /** @class */ (function (_super) {
-    __extends(HeatmapPoint, _super);
-    function HeatmapPoint() {
-        /* *
-         *
-         *  Properties
-         *
-         * */
-        var _this = _super !== null && _super.apply(this, arguments) || this;
-        _this.options = void 0;
-        _this.series = void 0;
-        _this.value = null;
-        _this.x = NaN;
-        _this.y = NaN;
-        return _this;
-        /* eslint-enable valid-jsdoc */
-    }
+
+    } as Highcharts.HeatmapSeriesOptions);
+
+    /* *
+     *
+     *  Properties
+     *
+     * */
+
+    public colorAxis: ColorAxis = void 0 as any;
+
+    public data: Array<HeatmapPoint> = void 0 as any;
+
+    public options: Highcharts.HeatmapSeriesOptions = void 0 as any;
+
+    public points: Array<HeatmapPoint> = void 0 as any;
+
+    public valueData?: Array<number>;
+
+    public valueMax: number = NaN;
+
+    public valueMin: number = NaN;
+
     /* *
      *
      *  Functions
      *
      * */
+
     /* eslint-disable valid-jsdoc */
+
     /**
      * @private
      */
-    HeatmapPoint.prototype.applyOptions = function (options, x) {
-        var point = LineSeries.prototype.pointClass.prototype
-            .applyOptions.call(this, options, x);
+    public drawPoints(): void {
+
+        // In styled mode, use CSS, otherwise the fill used in the style
+        // sheet will take precedence over the fill attribute.
+        var seriesMarkerOptions = this.options.marker || {};
+
+        if (seriesMarkerOptions.enabled || this._hasPointMarkers) {
+            LineSeries.prototype.drawPoints.call(this);
+            this.points.forEach((point): void => {
+                point.graphic &&
+                (point.graphic as any)[
+                    this.chart.styledMode ? 'css' : 'animate'
+                ](this.colorAttribs(point));
+            });
+        }
+    }
+
+    /**
+     * @private
+     */
+    getExtremes(): DataExtremesObject {
+        // Get the extremes from the value data
+        const { dataMin, dataMax } = LineSeries.prototype.getExtremes
+            .call(this, this.valueData);
+
+        if (isNumber(dataMin)) {
+            this.valueMin = dataMin;
+        }
+        if (isNumber(dataMax)) {
+            this.valueMax = dataMax;
+        }
+
+        // Get the extremes from the y data
+        return LineSeries.prototype.getExtremes.call(this);
+    }
+
+    /**
+     * Override to also allow null points, used when building the k-d-tree for
+     * tooltips in boost mode.
+     * @private
+     */
+    getValidPoints(
+        points?: Array<HeatmapPoint>,
+        insideOnly?: boolean
+    ): Array<Point> {
+        return LineSeries.prototype.getValidPoints.call(
+            this,
+            points,
+            insideOnly,
+            true
+        );
+    }
+
+    /**
+     * Define hasData function for non-cartesian series. Returns true if the
+     * series has points at all.
+     * @private
+     */
+    public hasData(): boolean {
+        return !!this.processedXData.length; // != 0
+    }
+
+    /**
+     * Override the init method to add point ranges on both axes.
+     * @private
+     */
+    public init(): void {
+        var options;
+
+        LineSeries.prototype.init.apply(this, arguments as any);
+
+        options = this.options;
+        // #3758, prevent resetting in setData
+        options.pointRange = pick(options.pointRange, options.colsize || 1);
+        // general point range
+        this.yAxis.axisPointRange = options.rowsize || 1;
+
+        // Bind new symbol names
+        extend(symbols, {
+            ellipse: symbols.circle,
+            rect: symbols.square
+        });
+    }
+
+    /**
+     * @private
+     */
+    public markerAttribs(
+        point: HeatmapPoint,
+        state?: string
+    ): SVGAttributes {
+        var pointMarkerOptions = point.marker || {},
+            seriesMarkerOptions = this.options.marker || {},
+            seriesStateOptions: PointStateHoverOptions,
+            pointStateOptions: PointStateHoverOptions,
+            shapeArgs = point.shapeArgs || {},
+            hasImage = point.hasImage,
+            attribs: SVGAttributes = {};
+
+        if (hasImage) {
+            return {
+                x: point.plotX,
+                y: point.plotY
+            };
+        }
+
+        // Setting width and height attributes on image does not affect
+        // on its dimensions.
+        if (state) {
+            seriesStateOptions = (seriesMarkerOptions as any).states[state] || {};
+            pointStateOptions = pointMarkerOptions.states &&
+                (pointMarkerOptions.states as any)[state] || {};
+
+            [['width', 'x'], ['height', 'y']].forEach(function (
+                dimension
+            ): void {
+                // Set new width and height basing on state options.
+                attribs[dimension[0]] = (
+                    (pointStateOptions as any)[dimension[0]] ||
+                    (seriesStateOptions as any)[dimension[0]] ||
+                    shapeArgs[dimension[0]]
+                ) + (
+                    (pointStateOptions as any)[dimension[0] + 'Plus'] ||
+                    (seriesStateOptions as any)[dimension[0] + 'Plus'] || 0
+                );
+
+                // Align marker by a new size.
+                attribs[dimension[1]] = shapeArgs[dimension[1]] +
+                    (shapeArgs[dimension[0]] - attribs[dimension[0]]) / 2;
+            });
+        }
+
+        return state ? attribs : shapeArgs;
+    }
+
+    /**
+     * @private
+     */
+    public pointAttribs(
+        point?: HeatmapPoint,
+        state?: StatesOptionsKey
+    ): SVGAttributes {
+        var series = this,
+            attr = LineSeries.prototype.pointAttribs.call(series, point, state),
+            seriesOptions = series.options || {},
+            plotOptions = series.chart.options.plotOptions || {},
+            seriesPlotOptions = plotOptions.series || {},
+            heatmapPlotOptions = plotOptions.heatmap || {},
+            stateOptions,
+            brightness,
+            // Get old properties in order to keep backward compatibility
+            borderColor =
+                seriesOptions.borderColor ||
+                heatmapPlotOptions.borderColor ||
+                seriesPlotOptions.borderColor,
+            borderWidth =
+                seriesOptions.borderWidth ||
+                heatmapPlotOptions.borderWidth ||
+                seriesPlotOptions.borderWidth ||
+                attr['stroke-width'];
+
+        // Apply lineColor, or set it to default series color.
+        attr.stroke = (
+            (point && point.marker && point.marker.lineColor) ||
+            (seriesOptions.marker && seriesOptions.marker.lineColor) ||
+            borderColor ||
+            this.color
+        );
+        // Apply old borderWidth property if exists.
+        attr['stroke-width'] = borderWidth;
+
+        if (state) {
+            stateOptions =
+                merge(
+                    (seriesOptions.states as any)[state],
+                    seriesOptions.marker &&
+                    (seriesOptions.marker.states as any)[state],
+                    point &&
+                    point.options.states &&
+                    (point.options.states as any)[state] || {}
+                );
+            brightness = stateOptions.brightness;
+
+            attr.fill =
+                stateOptions.color ||
+                H.color(attr.fill).brighten(brightness || 0).get();
+
+            attr.stroke = stateOptions.lineColor;
+        }
+
+        return attr;
+    }
+
+    /**
+     * @private
+     */
+    public setClip(animation?: (boolean|AnimationOptionsObject)): void {
+        var series = this,
+            chart = series.chart;
+
+        LineSeries.prototype.setClip.apply(series, arguments);
+        if (series.options.clip !== false || animation) {
+            (series.markerGroup as any)
+                .clip(
+                    (animation || series.clipBox) && series.sharedClipKey ?
+                        (chart as any)[series.sharedClipKey] :
+                        chart.clipRect
+                );
+        }
+    }
+
+    /**
+     * @private
+     */
+    public translate(): void {
+        var series = this,
+            options = series.options,
+            symbol = options.marker && options.marker.symbol || '',
+            shape = symbols[symbol] ? symbol : 'rect',
+            options = series.options,
+            hasRegularShape = ['circle', 'square'].indexOf(shape) !== -1;
+
+        series.generatePoints();
+        series.points.forEach(function (point): void {
+            var pointAttr,
+                sizeDiff,
+                hasImage,
+                cellAttr = point.getCellAttributes(),
+                shapeArgs = {
+                    x: Math.min(cellAttr.x1, cellAttr.x2),
+                    y: Math.min(cellAttr.y1, cellAttr.y2),
+                    width: Math.max(Math.abs(cellAttr.x2 - cellAttr.x1), 0),
+                    height: Math.max(Math.abs(cellAttr.y2 - cellAttr.y1), 0)
+                };
+
+            hasImage = point.hasImage =
+                (point.marker && point.marker.symbol || symbol || '')
+                    .indexOf('url') === 0;
+
+            // If marker shape is regular (symetric), find shorter
+            // cell's side.
+            if (hasRegularShape) {
+                sizeDiff = Math.abs(shapeArgs.width - shapeArgs.height);
+                shapeArgs.x = Math.min(cellAttr.x1, cellAttr.x2) +
+                    (shapeArgs.width < shapeArgs.height ? 0 : sizeDiff / 2);
+                shapeArgs.y = Math.min(cellAttr.y1, cellAttr.y2) +
+                    (shapeArgs.width < shapeArgs.height ? sizeDiff / 2 : 0);
+                shapeArgs.width = shapeArgs.height =
+                    Math.min(shapeArgs.width, shapeArgs.height);
+            }
+
+            pointAttr = {
+                plotX: (cellAttr.x1 + cellAttr.x2) / 2,
+                plotY: (cellAttr.y1 + cellAttr.y2) / 2,
+                clientX: (cellAttr.x1 + cellAttr.x2) / 2,
+                shapeType: 'path',
+                shapeArgs: merge(true, shapeArgs, {
+                    d: symbols[shape](
+                        shapeArgs.x,
+                        shapeArgs.y,
+                        shapeArgs.width,
+                        shapeArgs.height
+                    )
+                })
+            };
+
+            if (hasImage) {
+                point.marker = {
+                    width: shapeArgs.width,
+                    height: shapeArgs.height
+                };
+            }
+
+            extend(point, pointAttr);
+        });
+
+        fireEvent(series, 'afterTranslate');
+    }
+
+    /* eslint-enable valid-jsdoc */
+
+}
+
+/* *
+ *
+ *  Prototype Properties
+ *
+ * */
+
+interface HeatmapSeries {
+    alignDataLabel: typeof ColumnSeries.prototype.alignDataLabel;
+    axisTypes: typeof colorMapSeriesMixin.axisTypes;
+    colorAttribs: typeof colorMapSeriesMixin.colorAttribs;
+    colorKey: typeof colorMapSeriesMixin.colorKey;
+    drawLegendSymbol: typeof LegendSymbolMixin.drawRectangle;
+    getSymbol: typeof LineSeries.prototype.getSymbol;
+    parallelArrays: typeof colorMapSeriesMixin.parallelArrays;
+    pointArrayMap: Array<string>;
+    pointClass: typeof HeatmapPoint;
+    trackerGroups: typeof colorMapSeriesMixin.trackerGroups;
+}
+extend(HeatmapSeries.prototype, {
+
+    /**
+     * @private
+     */
+    alignDataLabel: ColumnSeries.prototype.alignDataLabel,
+
+    axisTypes: colorMapSeriesMixin.axisTypes,
+
+    colorAttribs: colorMapSeriesMixin.colorAttribs,
+
+    colorKey: colorMapSeriesMixin.colorKey,
+
+    directTouch: true,
+
+    /**
+     * @private
+     */
+    drawLegendSymbol: LegendSymbolMixin.drawRectangle,
+
+    hasPointSpecificOptions: true,
+
+    /**
+     * @ignore
+     * @deprecated
+     */
+    getBox: noop as any,
+
+    getExtremesFromAll: true,
+
+    getSymbol: LineSeries.prototype.getSymbol,
+
+    parallelArrays: colorMapSeriesMixin.parallelArrays,
+
+    pointArrayMap: ['y', 'value'],
+
+    trackerGroups: colorMapSeriesMixin.trackerGroups
+
+});
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+class HeatmapPoint extends ScatterSeries.prototype.pointClass {
+
+    /* *
+     *
+     *  Properties
+     *
+     * */
+
+    public options: Highcharts.HeatmapPointOptions = void 0 as any;
+
+    public pointPadding?: number;
+
+    public series: HeatmapSeries = void 0 as any;
+
+    public value: (number|null) = null;
+
+    public x: number = NaN;
+
+    public y: number = NaN;
+
+    /* *
+     *
+     *  Functions
+     *
+     * */
+
+    /* eslint-disable valid-jsdoc */
+
+    /**
+     * @private
+     */
+    public applyOptions(
+        options: Highcharts.HeatmapPointOptions,
+        x?: number
+    ): HeatmapPoint {
+        const point: HeatmapPoint = LineSeries.prototype.pointClass.prototype
+            .applyOptions.call(this, options, x) as any;
+
         point.formatPrefix =
             point.isNull || point.value === null ?
                 'null' : 'point';
+
         return point;
-    };
-    HeatmapPoint.prototype.getCellAttributes = function () {
-        var point = this, series = point.series, seriesOptions = series.options, xPad = (seriesOptions.colsize || 1) / 2, yPad = (seriesOptions.rowsize || 1) / 2, xAxis = series.xAxis, yAxis = series.yAxis, markerOptions = point.options.marker || series.options.marker, pointPlacement = series.pointPlacementToXValue(), // #7860
-        pointPadding = pick(point.pointPadding, seriesOptions.pointPadding, 0), cellAttr = {
-            x1: clamp(Math.round(xAxis.len -
-                (xAxis.translate(point.x - xPad, false, true, false, true, -pointPlacement) || 0)), -xAxis.len, 2 * xAxis.len),
-            x2: clamp(Math.round(xAxis.len -
-                (xAxis.translate(point.x + xPad, false, true, false, true, -pointPlacement) || 0)), -xAxis.len, 2 * xAxis.len),
-            y1: clamp(Math.round((yAxis.translate(point.y - yPad, false, true, false, true) || 0)), -yAxis.len, 2 * yAxis.len),
-            y2: clamp(Math.round((yAxis.translate(point.y + yPad, false, true, false, true) || 0)), -yAxis.len, 2 * yAxis.len)
-        };
+    }
+
+    public getCellAttributes(): Highcharts.HeatmapPointCellAttributes {
+        var point = this,
+            series = point.series,
+            seriesOptions = series.options,
+            xPad = (seriesOptions.colsize || 1) / 2,
+            yPad = (seriesOptions.rowsize || 1) / 2,
+            xAxis = series.xAxis,
+            yAxis = series.yAxis,
+            markerOptions = point.options.marker || series.options.marker,
+            pointPlacement = series.pointPlacementToXValue(), // #7860
+            pointPadding = pick(
+                point.pointPadding, seriesOptions.pointPadding, 0
+            ),
+            cellAttr: Highcharts.HeatmapPointCellAttributes = {
+                x1: clamp(Math.round(xAxis.len -
+                    (xAxis.translate(
+                        point.x - xPad,
+                        false,
+                        true,
+                        false,
+                        true,
+                        -pointPlacement
+                    ) || 0)
+                ), -xAxis.len, 2 * xAxis.len),
+
+                x2: clamp(Math.round(xAxis.len -
+                    (xAxis.translate(
+                        point.x + xPad,
+                        false,
+                        true,
+                        false,
+                        true,
+                        -pointPlacement
+                    ) || 0)
+                ), -xAxis.len, 2 * xAxis.len),
+
+                y1: clamp(Math.round(
+                    (yAxis.translate(
+                        point.y - yPad,
+                        false,
+                        true,
+                        false,
+                        true
+                    ) || 0)
+                ), -yAxis.len, 2 * yAxis.len),
+
+                y2: clamp(Math.round(
+                    (yAxis.translate(
+                        point.y + yPad,
+                        false,
+                        true,
+                        false,
+                        true
+                    ) || 0)
+                ), -yAxis.len, 2 * yAxis.len)
+            };
+
         // Handle marker's fixed width, and height values including border
         // and pointPadding while calculating cell attributes.
-        [['width', 'x'], ['height', 'y']].forEach(function (dimension) {
-            var prop = dimension[0], direction = dimension[1];
-            var start = direction + '1', end = direction + '2';
-            var side = Math.abs(cellAttr[start] - cellAttr[end]), borderWidth = markerOptions &&
-                markerOptions.lineWidth || 0, plotPos = Math.abs(cellAttr[start] + cellAttr[end]) / 2;
-            if (markerOptions[prop] &&
-                markerOptions[prop] < side) {
-                cellAttr[start] = plotPos - (markerOptions[prop] / 2) -
+        [['width', 'x'], ['height', 'y']].forEach(function (dimension): void {
+            const prop = dimension[0],
+                direction = dimension[1];
+
+            let start = direction + '1',
+                end = direction + '2';
+
+            const side = Math.abs(
+                    cellAttr[start] - cellAttr[end]
+                ),
+                borderWidth = markerOptions &&
+                    markerOptions.lineWidth || 0,
+                plotPos = Math.abs(
+                    cellAttr[start] + cellAttr[end]
+                ) / 2;
+
+
+            if (
+                (markerOptions as any)[prop] &&
+                (markerOptions as any)[prop] < side
+            ) {
+                cellAttr[start] = plotPos - (
+                    (markerOptions as any)[prop] / 2) -
                     (borderWidth / 2);
-                cellAttr[end] = plotPos + (markerOptions[prop] / 2) +
+
+                cellAttr[end] = plotPos + (
+                    (markerOptions as any)[prop] / 2) +
                     (borderWidth / 2);
             }
+
             // Handle pointPadding
             if (pointPadding) {
                 if (direction === 'y') {
@@ -627,88 +939,129 @@ var HeatmapPoint = /** @class */ (function (_super) {
                 cellAttr[end] -= pointPadding;
             }
         });
+
         return cellAttr;
-    };
+    }
+
     /**
      * @private
      */
-    HeatmapPoint.prototype.haloPath = function (size) {
+    public haloPath(size: number): SVGPath {
         if (!size) {
             return [];
         }
         var rect = this.shapeArgs;
+
         return [
             'M',
-            rect.x - size,
-            rect.y - size,
+            (rect as any).x - size,
+            (rect as any).y - size,
             'L',
-            rect.x - size,
-            rect.y + rect.height + size,
-            rect.x + rect.width + size,
-            rect.y + rect.height + size,
-            rect.x + rect.width + size,
-            rect.y - size,
+            (rect as any).x - size,
+            (rect as any).y + (rect as any).height + size,
+            (rect as any).x + (rect as any).width + size,
+            (rect as any).y + (rect as any).height + size,
+            (rect as any).x + (rect as any).width + size,
+            (rect as any).y - size,
             'Z'
         ];
-    };
+    }
+
     /**
      * Color points have a value option that determines whether or not it is
      * a null point
      * @private
      */
-    HeatmapPoint.prototype.isValid = function () {
+    public isValid(): boolean {
         // undefined is allowed
-        return (this.value !== Infinity &&
-            this.value !== -Infinity);
-    };
-    return HeatmapPoint;
-}(ScatterSeries.prototype.pointClass));
+        return (
+            this.value !== Infinity &&
+            this.value !== -Infinity
+        );
+    }
+
+    /* eslint-enable valid-jsdoc */
+
+}
+
+/* *
+ *
+ *  Prototype Properties
+ *
+ * */
+
+interface HeatmapPoint {
+    dataLabelOnNull: typeof colorMapPointMixin.dataLabelOnNull;
+    isValid: typeof colorMapPointMixin.isValid;
+    setState: typeof colorMapPointMixin.setState;
+}
 extend(HeatmapPoint.prototype, {
     dataLabelOnNull: colorMapPointMixin.dataLabelOnNull,
     setState: colorMapPointMixin.setState
 });
+
+/* *
+ *
+ *  Registry
+ *
+ * */
+
+declare module '../../Core/Series/SeriesType' {
+    interface SeriesTypeRegistry {
+        heatmap: typeof HeatmapSeries;
+    }
+}
 HeatmapSeries.prototype.pointClass = HeatmapPoint;
 BaseSeries.registerSeriesType('heatmap', HeatmapSeries);
+
 /* *
  *
  *  Default Export
  *
  * */
+
 export default HeatmapSeries;
+
 /* *
  *
  *  API Declarations
  *
  * */
+
 /**
  * Heatmap series only. Padding between the points in the heatmap.
  * @name Highcharts.Point#pointPadding
  * @type {number|undefined}
  */
+
 /**
  * Heatmap series only. The value of the point, resulting in a color
  * controled by options as set in the colorAxis configuration.
  * @name Highcharts.Point#value
  * @type {number|null|undefined}
  */
+
 /* *
  * @interface Highcharts.PointOptionsObject in parts/Point.ts
- */ /**
-* Heatmap series only. Point padding for a single point.
-* @name Highcharts.PointOptionsObject#pointPadding
-* @type {number|undefined}
-*/ /**
-* Heatmap series only. The value of the point, resulting in a color controled
-* by options as set in the colorAxis configuration.
-* @name Highcharts.PointOptionsObject#value
-* @type {number|null|undefined}
-*/
+ *//**
+ * Heatmap series only. Point padding for a single point.
+ * @name Highcharts.PointOptionsObject#pointPadding
+ * @type {number|undefined}
+ *//**
+ * Heatmap series only. The value of the point, resulting in a color controled
+ * by options as set in the colorAxis configuration.
+ * @name Highcharts.PointOptionsObject#value
+ * @type {number|null|undefined}
+ */
+
 ''; // detach doclets above
+
 /* *
  *
  *  API Options
  *
  * */
+
 /**
  * A `heatmap` series. If the [type](#series.heatmap.type) option is
  * not specified, it is inherited from [chart.type](#chart.type).
@@ -721,6 +1074,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap
  */
+
 /**
  * An array of data points for the series. For the `heatmap` series
  * type, points can be given in the following ways:
@@ -778,6 +1132,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data
  */
+
 /**
  * The color of the point. In heat maps the point color is rarely set
  * explicitly, as we use the color to denote the `value`. Options for
@@ -787,6 +1142,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.color
  */
+
 /**
  * The value of the point, resulting in a color controled by options
  * as set in the [colorAxis](#colorAxis) configuration.
@@ -795,6 +1151,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.value
  */
+
 /**
  * The x value of the point. For datetime axes,
  * the X value is the timestamp in milliseconds since 1970.
@@ -803,6 +1160,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.x
  */
+
 /**
  * The y value of the point.
  *
@@ -810,6 +1168,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.y
  */
+
 /**
  * Point padding for a single point.
  *
@@ -820,38 +1179,45 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.pointPadding
  */
+
 /**
  * @excluding radius, enabledThreshold
  * @product   highcharts highmaps
  * @since     8.1
  * @apioption series.heatmap.data.marker
  */
+
 /**
  * @excluding radius, enabledThreshold
  * @product   highcharts highmaps
  * @since     8.1
  * @apioption series.heatmap.marker
  */
+
 /**
  * @excluding radius, radiusPlus
  * @product   highcharts highmaps
  * @apioption series.heatmap.marker.states.hover
  */
+
 /**
  * @excluding radius
  * @product   highcharts highmaps
  * @apioption series.heatmap.marker.states.select
  */
+
 /**
  * @excluding radius, radiusPlus
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.marker.states.hover
  */
+
 /**
  * @excluding radius
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.marker.states.select
  */
+
 /**
 * Set the marker's fixed width on hover state.
 *
@@ -863,6 +1229,7 @@ export default HeatmapSeries;
 * @product   highcharts highmaps
 * @apioption series.heatmap.marker.states.hover.lineWidthPlus
 */
+
 /**
 * Set the marker's fixed width on hover state.
 *
@@ -874,6 +1241,7 @@ export default HeatmapSeries;
 * @product   highcharts highmaps
 * @apioption series.heatmap.marker.states.hover.width
 */
+
 /**
  * Set the marker's fixed height on hover state.
  *
@@ -885,6 +1253,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.marker.states.hover.height
  */
+
 /**
 * The number of pixels to increase the width of the
 * hovered point.
@@ -897,6 +1266,7 @@ export default HeatmapSeries;
 * @product   highcharts highmaps
 * @apioption series.heatmap.marker.states.hover.widthPlus
 */
+
 /**
  * The number of pixels to increase the height of the
  * hovered point.
@@ -909,6 +1279,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.marker.states.hover.heightPlus
  */
+
 /**
  * The number of pixels to increase the width of the
  * hovered point.
@@ -921,6 +1292,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.marker.states.select.widthPlus
  */
+
 /**
  * The number of pixels to increase the height of the
  * hovered point.
@@ -933,6 +1305,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.marker.states.select.heightPlus
  */
+
 /**
 * Set the marker's fixed width on hover state.
 *
@@ -944,6 +1317,7 @@ export default HeatmapSeries;
 * @product   highcharts highmaps
 * @apioption series.heatmap.data.marker.states.hover.lineWidthPlus
 */
+
 /**
  * Set the marker's fixed width on hover state.
  *
@@ -955,6 +1329,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.marker.states.hover.width
  */
+
 /**
  * Set the marker's fixed height on hover state.
  *
@@ -966,6 +1341,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.marker.states.hover.height
  */
+
 /**
  * The number of pixels to increase the width of the
  * hovered point.
@@ -978,6 +1354,7 @@ export default HeatmapSeries;
  * @product   highcharts highstock
  * @apioption series.heatmap.data.marker.states.hover.widthPlus
  */
+
 /**
  * The number of pixels to increase the height of the
  * hovered point.
@@ -990,6 +1367,7 @@ export default HeatmapSeries;
  * @product   highcharts highstock
  * @apioption series.heatmap.data.marker.states.hover.heightPlus
  */
+
 /**
 * Set the marker's fixed width on select state.
 *
@@ -1001,6 +1379,7 @@ export default HeatmapSeries;
 * @product   highcharts highmaps
 * @apioption series.heatmap.data.marker.states.select.width
 */
+
 /**
  * Set the marker's fixed height on select state.
  *
@@ -1012,6 +1391,7 @@ export default HeatmapSeries;
  * @product   highcharts highmaps
  * @apioption series.heatmap.data.marker.states.select.height
  */
+
 /**
  * The number of pixels to increase the width of the
  * hovered point.
@@ -1024,6 +1404,7 @@ export default HeatmapSeries;
  * @product   highcharts highstock
  * @apioption series.heatmap.data.marker.states.select.widthPlus
  */
+
 /**
  * The number of pixels to increase the height of the
  * hovered point.
@@ -1036,4 +1417,5 @@ export default HeatmapSeries;
  * @product   highcharts highstock
  * @apioption series.heatmap.data.marker.states.select.heightPlus
  */
+
 ''; // adds doclets above to transpiled file
